v0.3.0 - DB connection and other changes
Database connection is now used in each and every file where database transactions are required, unlike previous versions where a single file had the database connection. This is done to get latest values by opening and closing connections.

"Report" button text is changed to "Export to XLSX" and a new icon for it has been added.

PS: The previous version was 0.2.0 and not 2.0.0. This one is v0.3.0.